+++ src/main/java/com/aac/backend/presentation/dto/response/SentenceResponse.java
@@ -1,3 +1,3 @@
-package com.aac.backend.controller.dto.response;
+package com.aac.backend.presentation.dto.response;
 
 public record SentenceResponse(String sentence) {}
